+++ CHANGELOG.rst
@@ -1,3 +1,16 @@
+Version 0.2.0:
+  * Option to turn off tick label rotation
+  * Long tick labels are now lined up correctly with their ticks
+  * Option to fine tune position of label relative to tick
+  * Can choose to use 1d or 2d Gaussian sigma contours
+  * scipy is now an optional dependency
+  * Can choose to fill contours and 1d histograms independently
+  * Can plot the actual data-point density underneath the contours
+  * Full control over fonts for labels, ticks, and legend
+  * Use built-in matplotlib LaTex renderer for equations in labels so no LaTex
+    installation is required locally.
+  * Added a new example to the documentation
+
 Version 0.1.1:
   * Prettier truth colors
   * Parameter labels line up even if parameter numbers are very different in size